Rc (#25)
* new helper scdriupt added

* Add files via upload

* Delete skripts/export/common.conf

* Delete skripts/export/common_funcs.sh

* Delete skripts/export/loits2.sh

* few more pkgs added

* cmdline parsing maybe works

* ... works with loits_new i mean

* enforce_deps kinda works

* inst_dep kinda wks, s0rt 0f

* Add files via upload

* sudo-related

* still trying 2 ake loits_new work

* grub-mkrescue

* Add files via upload

* fixing sme things

* trying 2 make loits2 w0rk

* loits2 opt10n pars1ng

* l2.check.pars.1

* ls2cp3

* sco,scm

* debug added 2 loits2

* make_tar()

* mostly abt mkconffile.sh

* Delete skripts/export/mkconffile.sh.export

* Add files via upload

* new file

* Delete stage0.sh.export

* new file

* Delete skripts/export/stage0_backend.bsh.export

* try1ng 2 mak3 backend w0rk again

* mostly abt commments

* comments2

* mostly abt make_tgt_dirs()

* mostly abt make_src_dirs()

* trying 2 guess what i meant by "make_dirs"

* mostly abt "sudo cp"

* mostöyu abt "sudo rm"

* Add files via upload

* Delete stage0f.sh.export

* debug

* mostly abt sudo

* new stuff

* new stuff 2

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* i hope these scripts are in "almost works" state so that

... i can soon test with an actual .iso-file

* i hope...

* vommentz

* Update common_funcs.sh

vommentz abt future devel

* trying 2 make stage0+stage0f work

* trying 2 make stage0+stage0f work

* Add files via upload

* Add files via upload

* maybe bootloader()

* maybe bootloader() works

... sho that i can concentrate on "loits"-scripts

* changing how to initialize tha "dev_env"

* conf_gi_opts changed for loits_skripts

* changes

* changes.2

i hope that loits_new starts to work soon

* init-dev-env-related

* saturday's changes

* saturday's changes 2

- i hope that loits_new's output boots now
(not yet 100% because Reasons)

- also some changes to "init-scripts"

* tried 2 mak3 loits2 work

Not exactly sure why there are 2 different scripts for same job. 
Was there an actual reason 4 that or not? Maybe i was just bored and decided to write a new version.

* dependencies

* sunday's changes 2

* sunday's changes 3

* commentsa

* cleanup

* cleanup

* maybe

* deps

actually i'm supposed to make new .iso-files boot, instead stick on doing sometrging else

* deps again

* grub almost works

actually  tonighti made a grub-bootable .iso once, by accident
next;make grub-bootable happen on purpose

* grub almost works+other changes

* pkg mtools retunred

* monday's changes 1

* monday's changes 2

* Delete skripts/export/mksums_new.sh.export

* Delete skripts/export/common_funcs.sh.export

* Delete skripts/export/loits2.sh.export

* Delete skripts/export/loits_new.sh

* Delete skripts/export/loits_new.sh.export

* Add files via upload

* maybe mksums kinda?

* friday's changes 1

* pkg install changed a bit

* new script added

may not tey work properly

* Delete skripts/export/squ.ash.export

* mostly abt mksums

* freya's dag 2

* changes here and there

maybe mksums works properly now?
removed debug
changed grub-spesific stuff

+++ skripts/init2.sh
@@ -22,14 +22,16 @@ if [ -s ${basedir}/interfaces ] ; then
 	sudo ln -s /etc/network/interfaces.${distro} /etc/network/interfaces
 fi
 
-c=$(ls ${pkgsrc}/ip*.deb | wc -l)
+#tarpeellinen kikkailu?
+#c=$(ls ${pkgsrc}/ip*.deb | wc -l)
 
 function efk() {
 	sudo dpkg -i $@
 	sudo rm $@
 }
 
-if [ ${c} -gt 0 ] ; then
+#völillä jos kokeilusi ajaa init2 ennen generic_x- juttui
+#if [ ${c} -gt 0 ] ; then
 	q=$(mktemp -d)
 	sudo cp ${pkgsrc}/*.deb ${q}
 	
@@ -40,49 +42,70 @@ if [ ${c} -gt 0 ] ; then
 	sudo dpkg -i ${q}/lib*.deb
 	sudo rm ${q}/lib*.deb
 
+	#HUOM.17725:josqo lib-juttujen jälkeen pak as vain tarvittaessa, which...
 	echo "BEFORE TBLZ"
 	sleep 2
 
 	#DEBIAN_FRONTEND=noninteractive tämän kanssa jokin juttu?
-	sudo dpkg -i ${q}/iptables_*.deb
-	sudo rm ${q}/iptables_*.deb
 
-	sudo dpkg -i ${q}/net*.deb
-	sudo rm ${q}/net*.deb
+	#if_not_iptables
+	x=$(sudo which iptables)
 
-	#tähän kai uskaltaa laittaa frontend takaisinb
-	sudo dpkg -i ${q}/iptables-*.deb
-	sudo rm ${q}/iptables-*.deb
+	if [ ! -x ${x} ] ; then
+		sudoDEBIAN_FRONTEND=noninteractive dpkg -i ${q}/iptables_*.deb
+		sudo rm ${q}/iptables_*.deb
 
-	#deb uig taakse jatqssa
-	echo "after tables"
-	dpkg -l iptables*
-	sleep 2
+		sudo dpkg -i ${q}/net*.deb
+		sudo rm ${q}/net*.deb
+
+		#tähän kai uskaltaa laittaa frontend takaisinb
+		sudo DEBIAN_FRONTEND=noninteractive dpkg -i ${q}/iptables-*.deb
+		sudo rm ${q}/iptables-*.deb
+
+		#deb uig taakse jatqssa
+		echo "after tables"
+		dpkg -l iptables*
+		sleep 2
+	fi
+	#/if_not_iptables
 
 	#qseeko tässä kohtaa jokin?
+	#if_not_git
+	x=$(sudo which git)
 
-	sudo dpkg -i $q/git-man*.deb
-	sudo rm ${q}/git-man*.deb
+	if [ ! -x ${x} ] ; then
+		sudo dpkg -i $q/git-man*.deb
+		sudo rm ${q}/git-man*.deb
 
-	sudo dpkg -i $q/git*.deb
-	sudo rm ${q}/git*.deb
+		sudo dpkg -i $q/git*.deb
+		sudo rm ${q}/git*.deb
+	fi
+	#/if_not_git
 
-	sudo dpkg -i $q/grub*.deb
-	sudo rm ${q}/grub*.deb
+	#if_not_grub
+	x=$(sudo which grub-mkrescue)
+
+	if [ ! -x ${x} ] ; then
+		sudo dpkg -i $q/grub*.deb
+		sudo rm ${q}/grub*.deb
+	fi
+	#/if_not_grub
 
 	#ao. rivejä ei kannata unmohtaa
+	#which grub geniso
+	#which grub xorrisao...
 	sudo dpkg -i $q/*.deb
 	sudo rm ${q}/*.deb
-else
-	sudo dpkg -i ${pkgsrc}/*.deb
-fi
+#else
+#	sudo dpkg -i ${pkgsrc}/*.deb
+#fi
 
 echo "GENISOIMAGE?"
 which genisoimage
 sleep 6
 
 #pois kommenteista 14725, takaisin jos qsee
-sudo apt-get remove --purge python3-cups ntp*
+sudo apt-get remove --purge --yes python3-cups ntp* #sharyp from common_lib
 sudo apt autoremove
 
 sudo which iptables-restore
